Rename baseTest into test-utils
