+++ example/ios/Supporting Files/settings.gradle.kts
@@ -32,8 +32,8 @@ apply("../../../gradle/ext.gradle")
 val ext =  (gradle as ExtensionAware).extra
 
 if (!(ext["exampleAsRoot"] as Boolean)) {
-    include(":baseTest")
-    project(":baseTest").projectDir = file("../../../baseTest")
+    include(":test-utils")
+    project(":test-utils").projectDir = file("../../../test-utils")
 
     include(":location")
     project(":location").projectDir = file("../../../location")
