+++ settings.gradle
@@ -31,7 +31,7 @@ rootProject.name = 'Kaluga'
 apply from: "gradle/ext.gradle"
 
 include ':base'
-include ':baseTest'
+include ':test-utils'
 include ':alerts'
 include ':logging'
 include ':loadingIndicator'
