+++ test-utils/src/androidLibAndroidTest/AndroidManifest.xml
@@ -1,3 +1,3 @@
-<manifest package="com.splendo.kaluga.basetest"
+<manifest package="com.splendo.kaluga.test"
           xmlns:android="http://schemas.android.com/apk/res/android"
 />
