+++ test-utils/src/iosMain/kotlin/test/GlobalTestListener.kt
@@ -1,4 +1,4 @@
-package com.splendo.kaluga.basetest
+package com.splendo.kaluga.test
 /*
 
 Copyright 2019 Splendo Consulting B.V. The Netherlands
@@ -17,29 +17,11 @@ Copyright 2019 Splendo Consulting B.V. The Netherlands
 
 */
 
-import kotlin.test.AfterTest
-import kotlin.test.BeforeTest
-
-/*
- * Android Studio at this time thinks this class has too many implementations because of the test and androidTest source sets.
- * To clear the error close the file and restart Android Studio ¯\_(ツ)_/¯
- */
-expect class GlobalTestListener() {
-    fun beforeTest()
-    fun afterTest()
-}
-
-open class BaseTest {
-    private val testListener = GlobalTestListener()
-
-    @BeforeTest
-    fun beforeTest() {
-        testListener.beforeTest()
+actual class GlobalTestListener {
+    actual fun beforeTest() {
     }
 
-    @AfterTest
-    fun afterTest() {
-        testListener.afterTest()
+    actual fun afterTest() {
     }
 
 }
